Biginner Level/requirements.txt
streamlit
tensorflow
numpy
pillow
